requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.6
django-appconf==1.0.2
django-bootstrap-datepicker-plus==3.0.4
django-bootstrap-modal-forms==1.3.1
django-bootstrap3==11.0.0
django-braces==1.13.0
django-ckeditor==5.6.1
django-colorful==1.3
django-crispy-forms==1.7.2
django-el-pagination==3.2.4
django-embed-video==1.2.0
django-import-export==1.1.0
django-js-asset==1.1.0
django-mathfilters==0.4.0
django-model-utils==3.1.2
django-modelcluster==4.2
django-notifications-hq==1.5.0
django-pagedown==1.0.6
django-ranged-response==0.2.0
django-select2==6.3.1
django-simple-captcha==0.5.9
django-social-share==1.3.0
django-sortedm2m==1.5.0
django-sslserver==0.20
django-star-ratings==0.7.0
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.0.0
et-xmlfile==1.0.1
Faker==0.8.16
html5lib==1.0.1
idna==2.7
jdcal==1.4
jsonfield==2.0.2
lml==0.0.7
numpy==1.15.4
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.12
pandas==0.23.4
Pillow==5.2.0
pycparser==2.18
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
PyJWT==1.6.4
PyPDF2==1.26.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.13
reportlab==3.5.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sorl-thumbnail==12.5.0
swapper==1.1.0
tablib==0.12.1
text-unidecode==1.2
texttable==1.5.0
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.23
webencodings==0.5.1
Willow==1.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
